# rgw/s3-notifications: async commit to persistent notification queue