@yuvalif 2024-01-21
my main concern here is the case the the change to the queue is slower than the actual rados operations trigerring the notifications.
when `publish_commit()` was syncronouse it was just slowing down the operations, but with this change, the completion objects will just build up and eventually take all memeory.
can we add a counter for pending publish commit completions, and reply with "busy" error to the client if the counter crosses a threshold?