@igomon-bloomberg 2024-01-22
I see. Do you have a suggestion of how we could define the threshold value?