@igomon-bloomberg 2024-01-22
> > my main concern here is the case the the change to the queue is slower than the actual rados operations trigerring the notifications. when `publish_commit()` was syncronouse it was just slowing down the operations, but with this change, the completion objects will just build up and eventually take all memeory. can we add a counter for pending publish commit completions, and reply with "busy" error to the client if the counter crosses a threshold?
> 
> had a discussion about that with @cbodley, the number of pending completions will not exceed the number of pending reservations. the `publish_reserve()` command is syncronouse with the opration and would fail the operation onc we ran out of space in the queue, even if this space is only reserved and not actually used.

Ok, I see. So we are then good to go without the threshold check?