@yuvalif 2024-01-23
* please squash commits
* please run it against the `rgw:notifications` teuthology suite