@yuvalif 2024-01-25
> @yuvalif @mkogan1 this should be a nice performance boost. do you think it'd be worth trying to measure that?

Yes. This isnthe plan once the code is finalized