@yuvalif 2024-01-28
perf test instructions are here: https://gist.github.com/yuvalif/5587af49ebed90d2bed10d3312223e5c

> could not find a difference between a difference in the drop of performence between "main" and this branch. in both cases IO/s drop from ~1200 to ~750