@mkogan1 2024-01-28
> perf test instructions are here: https://gist.github.com/yuvalif/5587af49ebed90d2bed10d3312223e5c
> 
> > could not find a difference between a difference in the drop of performence between "main" and this branch. in both cases IO/s drop from ~1200 to ~750 for `PUT` and from ~1300 to ~850 for `DEL`

although there is no perceived difference in `hsbench` workload performance between this PR and main 
(it is probably capped by disk IO utilization)
few things that can think of - if we also compare the before and after this PR of
- latency (hsbench `Lat(ms): [ min: ..., avg: ..., 99%: ..., max: ...]` )
- radosgw process CPU utilization

one of those metrics could show an advantage of this commit  
(myb there are other notification related matrices worth evaluating)