@cbodley 2024-01-28
> yes, if no events are specified, we notify on all creation and deletion events

i see. is that special behavior for rgw? i didn't see anything like that from aws docs