@yuvalif 2024-01-28
# CPU Utilization (baseline)
![cpu-main](https://github.com/ceph/ceph/assets/1798418/7f9fad26-0a92-4c6c-8481-b6fbd1805aed)
# hsbench Average Latency (ms) and  IO/s (baseline)
![hsbench-main](https://github.com/ceph/ceph/assets/1798418/34ebfcc1-6c27-43c4-93cc-96c24cf661e8)