@yuvalif 2024-01-28
# CPU Utilization (this PR):
![cpu1](https://github.com/ceph/ceph/assets/1798418/7156861a-0e89-4b8c-a361-8bf203612c63)
# hsbench Average Latency (ms) and IO/s (this PR):
![hsbench-async](https://github.com/ceph/ceph/assets/1798418/fc9e5f7b-2695-486b-8ad9-3450f76fddb7)