@igomon-bloomberg 2024-01-29
That is surprising. I did see clear reduction in latencies and increased PUT/s in my local testing. Overall, the penalty for enabling bucket notification was up to 40% less for object sizes up to 4K with this PR. I used warm put benchmark running against a VM cluster. We are also going to do performance testing against a cluster running on a bare metal machines soon.