@mkogan1 2024-01-29
@yuvalif, regarding why performance dosent scale above 2K IOPS , it is because 
```
hsbench -a 0555b35654ad1656d804 -s h7GhxuBLTrlhVUyxSPUKUV8r/2EI4ngqJxD7iBdBYLhwluN30JaT3Q== \
  -u http://localhost:8000/ -bp bk -m ipd -t 64 -z 1K -l 100 -ri 10 -d 120
 ```
is operation on a single bucket, lets please try to increase the number of buckets, ex add  `-b 6` to see if it scales up
(6*2000 = 12000 IOPS, this system has performed 11683 IOPS in my hsbench 4KB PUT test run)