@mkogan1 2024-01-30
updating for future reference, worked with @yuvalif on slack and we found that the PGs number had to be increased to get nominal performance:
`-o osd_pool_default_pg_num=128 -o osd_pool_default_pgp_num=128 -o mon_max_pg_per_osd=32768 -o mon_pg_warn_max_per_osd=32768 -o osd_pool_default_pg_autoscale_mode=warn`