@yuvalif 2024-01-30
> > yes, if no events are specified, we notify on all creation and deletion events
> 
> i see. is that special behavior for rgw? i didn't see anything like that from aws docs

@cbodley you are right, but this is the way we configure that in the past 4 years, so probably not going to change that...