@igomon-bloomberg 2024-01-30
> > That is surprising. I did see clear reduction in latencies and increased PUT/s in my local testing. Overall, the penalty for enabling bucket notification was up to 40% less for object sizes up to 4K with this PR. I used warp put benchmark running against a VM cluster. We are also going to do performance testing against a cluster running on a bare metal machines soon. My measurement is based on the latency attribute of the PUT op response as reported in Ceph's logs.
> 
> @igomon-bloomberg could you provide more details on how to reproduce your results?

@yuvalif my test setup is a single-site VM cluster (2 RGWs, 3 MONs, 3 OSDs). The command I used was:
warp put --obj.size=<1K/4K> --access-key=.. --secret-key=... --host=.. --bucket=test-bucket
The notifications were sent to a Kafka endpoint.
I calculated average latency over all PUT requests as reported by latency attribute value in Ceph logs. 
The results I got for my single-site cluster are:

WITHOUT PR:
Obj Size |  Notifications On | Notifications Off
---------------------------------------------
256        |  54ms                   | 28ms
1K           |  56ms                   | 29ms
4K          |  57ms                    | 30ms

WITH PR:
Obj Size |  Notifications On | Notifications Off
---------------------------------------------
256        |  45ms                   | 32ms
1K           |  47ms                   | 29ms
4K          |  50ms                    | 31ms

However I did more tests recently (against a multisite cluster) and I noticed that latencies jump much higher when notifications are enabled. I only recorded numbers for 4K:
WITHOUT PR:
Obj Size |  Notifications On | Notifications Off
---------------------------------------------
4K       |  188ms                   | 32ms

WITH PR:
Obj Size |  Notifications On | Notifications Off
---------------------------------------------
4K       |  187ms                   | 29ms

Not quite what I was expecting and seems to me like maybe less CPU/disk capacity is available due to multisite processing?